+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 BUG FIXES:
 
-  - provisioners/ansible_local: Fix error in `playbook` existence check when
-    running on a Windows host [GH-6740]
+  - provisioners/ansible_local: Fix errors in absolute paths to playbook or
+      galaxy resources when running on a Windows host [GH-6740, GH-6757]
+  - provisioners/ansible_local: Change the way to verify `ansible-galaxy`
+      presence, to avoid a non-zero status code with Ansible 2.0 [GH-6793]
 
 ## 1.8.1 (December 21, 2015)